ptbcontrib/postgres_persistence/requirements.txt
python-telegram-bot>=12.0
SQLAlchemy
